Modified the Docker file to run tests inside the image.
Using a Docker multi-stage image to build stages to run unit tests, generate test reports, and an image for deployments. This allows the image we push to AWS ECR to be about 50% smaller (520MB vs 240MB) than with the previous configuration. This is because git other non-core tools are not installed in that image.