Integrating CodeDeploy into CodePipeline for Lambda
It seems CodeDeploy cannot retrieve AppSpec.yaml file from the artifact. So, we use CodeBuild output artifacts to pipe it into CodeDeploy.

Also, AppSpec.yaml requires that we specifiy the current and target version. This does not quite make sense because, given an alias, we should be able to compute the current version.

In the end, we decided to create a function to create the AppSpec.yaml file.